+++ lib/src/core/resolver.dart
@@ -1,0 +1,5 @@
+//Copyright (C) 2013 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+part of core;
+